\name{wines}
\alias{wines}
\alias{wine.classes}
\docType{data}
\title{Wine data}
\description{
  A data frame containing 177 rows and thirteen columns; object
  \code{wine.classes} contains the class labels.
  
  These data are the results of chemical analyses of
  wines grown in the same region in Italy but derived from three
  different cultivars. The data contain the quantities of several
  constituents found in each of the three types of wines, as well as
  some spectroscopic variables.
}
\usage{data(wines)}
\source{
  \url{http://kdd.ics.uci.edu}
}
\keyword{datasets}
