\name{yeast}
\alias{yeast}
\docType{data}
\title{Yeast cell-cycle data}
\description{
  Microarray cell-cycle data for 800 yeast genes, arrested with six
  different methods, arranged in a list. Additional class information is
  present as well.
}
\usage{data(yeast)}
\references{P. Spellman et al., Mol. Biol. Cell 9, 3273-3297 (1998)}
\keyword{datasets}
