\name{wines}
\alias{wines}
\alias{wine.classes}
\docType{data}
\title{Wine data}
\description{
  These data are the results of a chemical analysis of
  wines grown in the same region in Italy but derived from three
  different cultivars. The analysis determined the quantities of 13
  constituents found in each of the three types of wines.
}
\usage{data(wines)}
\source{
  \url{http://kdd.ics.uci.edu}
}
\keyword{datasets}
