\name{wines}
\alias{wines}
\alias{vintages}
\docType{data}
\title{Wine data}
\description{
  A data frame containing 177 rows and thirteen columns; object
  \code{vintages} contains the class labels. 
  
  These data are the results of chemical analyses of
  wines grown in the same region in Italy (Piedmont) but derived from
  three different cultivars: Nebbiolo, Barberas and Grignolino grapes. The
  wine from the Nebbiolo grape is called Barolo.
  The data contain the quantities of several
  constituents found in each of the three types of wines, as well as
  some spectroscopic variables.
}
\usage{data(wines)}
\references{M. Forina, C. Armanino, M. Castino and M. Ubigli. Vitis,
  25:189-201 (1986)} 
\source{
  \url{http://kdd.ics.uci.edu}
}
\keyword{datasets}
